Packages renamed to plural versions to match the rest
