+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/steps/VariantsAnnotLoad.java
@@ -31,7 +31,7 @@ import org.springframework.data.mongodb.core.MongoOperations;
 import uk.ac.ebi.eva.pipeline.configuration.VariantJobsArgs;
 import uk.ac.ebi.eva.pipeline.io.readers.VariantAnnotationReader;
 import uk.ac.ebi.eva.pipeline.io.writers.VariantAnnotationMongoItemWriter;
-import uk.ac.ebi.eva.pipeline.listener.SkipCheckingListener;
+import uk.ac.ebi.eva.pipeline.listeners.SkipCheckingListener;
 import uk.ac.ebi.eva.utils.MongoDBHelper;
 
 import java.io.IOException;
