+++ src/main/java/uk/ac/ebi/eva/pipeline/listeners/SkipCheckingListener.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.ac.ebi.eva.pipeline.listener;
+package uk.ac.ebi.eva.pipeline.listeners;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
